chore: add explicit cast

+++ sensing/autoware_pointcloud_preprocessor/src/downsample_filter/random_downsample_filter_node.cpp
@@ -58,7 +58,7 @@ RandomDownsampleFilterComponent::RandomDownsampleFilterComponent(
 {
   // set initial parameters
   {
-    sample_num_ = declare_parameter<int64_t>("sample_num");
+    sample_num_ = static_cast<size_t>(declare_parameter<int64_t>("sample_num"));
   }
 
   using std::placeholders::_1;
